Move files into utils folder and delete test file
