--- test.py
@@ -1,3 +1,0 @@
-import github_dates
-
-print(github_dates.current_year())
